Move input element + catch mission set name on first render
