+++ src/web/components/MissionsPanel/MissionsPanel.less
@@ -4,3 +4,10 @@
     row-gap: 15px;
     width: 300px;
 }
+
+.mission-set-name {
+    all: unset;
+    border: 1px solid #767d7d;
+    color: #fff;
+    padding: 3px 6px;
+}
