+++ src/web/components/MissionsPanel/MissionsPanel.tsx
@@ -6,7 +6,7 @@ import MissionsList from "./MissionsList/MissionsList";
 import DeleteMissionButton from "../../components/__buttons__/DeleteMissionButton/DeleteMissionButton";
 import MissionSpeedSliders from "./MissionSpeedSliders/MissionSpeedSliders";
 import MissionSetStorageButton from "./MissionSetStorage/MissionSetStorageButton";
-import { JaiaDispatchContext } from "../../context/JaiaContext";
+import { JaiaContext, JaiaDispatchContext } from "../../context/JaiaContext";
 import { JaiaActions } from "../../context/jaia-actions";
 import { MDI_BUTTON_SIZE } from "../../utils/constants";
 import { scrollMissionsList } from "../../utils/style";
@@ -18,10 +18,17 @@ import { mdiAutoFix, mdiPlus } from "@mdi/js";
 
 import "./MissionsPanel.less";
 
+interface Props {
+    missionSetName?: string;
+    numOfMissions?: number;
+    handleNameChange?: (name: string) => void;
+}
+
 /**
  * Renders a panel for operators to manage missions
  */
 export default function MissionsPanel() {
+    const jaiaContext = useContext(JaiaContext);
     const jaiaDispatch = useContext(JaiaDispatchContext);
 
     /**
@@ -43,6 +50,19 @@ export default function MissionsPanel() {
         jaiaDispatch({ type: JaiaActions.AUTO_ASSIGN_MISSIONS });
     };
 
+    /**
+     * Dispatches action to update the mission set name
+     *
+     * @param {string} name Input entered by operator
+     * @returns {void}
+     */
+    const handleNameChange = (name: string) => {
+        jaiaDispatch({
+            type: JaiaActions.CHANGE_MISSION_SET_NAME,
+            missionSetName: name,
+        });
+    };
+
     return (
         <div className="jaia-panel missions-panel">
             <div className="jaia-panel-title">Mission Set</div>
@@ -56,7 +76,7 @@ export default function MissionsPanel() {
                     <Icon path={mdiPlus} size={MDI_BUTTON_SIZE} title="Add mission" />
                 </Button>
                 <DeleteMissionButton deleteAll={true} />
-                <MissionSetStorageButton />
+                <MissionSetStorageButton missionSetName={jaiaContext.missionSet.getName()} />
                 <Button
                     className="jaia-button"
                     aria-label="auto-assign-bots"
@@ -65,7 +85,24 @@ export default function MissionsPanel() {
                     <Icon path={mdiAutoFix} size={MDI_BUTTON_SIZE} title="Auto assign Bots" />
                 </Button>
             </div>
+            <MissionSetName
+                missionSetName={jaiaContext.missionSet.getName()}
+                numOfMissions={jaiaContext.missionSet.getMissions().size}
+                handleNameChange={handleNameChange}
+            />
             <MissionsList />
         </div>
     );
+}
+
+function MissionSetName(props: Props) {
+    if (props.numOfMissions > 0) {
+        return (
+            <input
+                className="mission-set-name"
+                value={props.missionSetName}
+                onChange={(event) => props.handleNameChange(event.target.value)}
+            />
+        );
+    }
 }
